keys in json template are case insensitive. Basically converted to lowercase
