+++ config.py
@@ -64,8 +64,6 @@ CLR_GRAY = (130, 130, 130)
 # CLR_DARK_GRAY = (190,190,190)
 CLR_DARK_GRAY = (100, 100, 100)
 
-MIN_PAGE_AREA = 80000
-
 # Filepaths
 class Paths:
     def __init__(self, output):
